+++ HISTORY.rst
@@ -4,6 +4,10 @@ History
 Next Release
 ------------
 
+* Add test for identifying purely metabolic reactions with fixed constraints in
+  models
+* Add test for identifying transport reactions with fixed constraints in models
+
 0.6.2 (2018-03-12)
 ------------------